Create separate component for `TimeRangePickerTabs`.
